sequencer: honor signoff opt in run_git_commit
When rebasing interactively, --signoff would not take effect for commits
which conflict. That is, commits applying cleanly would be signed off,
but commits requiring intervention would miss the sign off trailer.

The reason is that run_git_commit did not check for the signoff replay
opt, and hence even though the option was picked up and passed
correctly, the actual committing dropped the ball.

The patch adds a test specifically for this case, as well as amending a
squash test which codified the broken behaviour.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>